backend/requirements.txt
flask
flask-cors
flask-mail
bcrypt
pycryptodome
web3
py-solc-x
gunicorn